+++ src/org/nolat/rsircbot/tools/URLReader.java
@@ -3,6 +3,7 @@ package org.nolat.rsircbot.tools;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.URL;
+import java.net.URLConnection;
 
 public class URLReader {
 
@@ -11,12 +12,16 @@ public class URLReader {
         StringBuffer buffer = new StringBuffer();
         try {
             URL url = new URL(urlString);
-            reader = new BufferedReader(new InputStreamReader(url.openStream()));
+            URLConnection connection = url.openConnection();
+            connection.setRequestProperty("User-Agent", "Mozilla/5.0 (Windows NT 6.1; WOW64) " +
+                    "AppleWebKit/537.11 (KHTML, like Gecko) " +
+                    "Chrome/23.0.1271.95 " +
+                    "Safari/537.11");
 
-            int read;
-            char[] chars = new char[1024];
-            while ((read = reader.read(chars)) != -1) {
-                buffer.append(chars, 0, read);
+            reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+            String line;
+            while ((line = reader.readLine()) != null) {
+                buffer.append(line);
             }
 
         } finally {